requirements.txt
torch
torchvision
matplotlib
numpy
pillow
scikit-learn
seaborn
notebook
